reorganize sqls, optimize range query for sqlite
