+++ connectorx/src/lib.rs
@@ -16,6 +16,7 @@ pub mod dummy_typesystem;
 pub mod errors;
 pub mod partition;
 pub mod sources;
+pub mod sql;
 pub mod transports;
 
 pub use crate::data_order::DataOrder;
